requirements.txt
Django==3.1.4
Pillow==8.0.1
django-admin-sortable2==0.7.7